VZ-7360: Refactor package name and locations
