+++ platform-operator/apis/verrazzano/webhooks/mysql_backup_webhook_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package mysql_webhook
+package webhooks
 
 import (
 	"context"
